+++ sei-ibc-go/modules/light-clients/07-tendermint/types/misbehaviour_test.go
@@ -26,7 +26,6 @@ func (suite *TendermintTestSuite) TestMisbehaviour() {
 
 	suite.Require().Equal(exported.Tendermint, misbehaviour.ClientType())
 	suite.Require().Equal(clientID, misbehaviour.GetClientID())
-	suite.Require().Equal(height, misbehaviour.GetHeight())
 }
 
 func (suite *TendermintTestSuite) TestMisbehaviourValidateBasic() {